Move VTK-only code to VTK module
