--- tensorvis/tensorvisio/include/modules/tensorvisio/processors/vtkunstructuredgridtorectilineargrid.h
@@ -1,99 +1,0 @@
-/*********************************************************************************
- *
- * Inviwo - Interactive Visualization Workshop
- *
- * Copyright (c) 2017-2018 Inviwo Foundation
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- * list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following disclaimer in the documentation
- * and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- *********************************************************************************/
-
-#ifndef IVW_VTKUNSTRUCTUREDGRIDTORECTILINEARGRID_H
-#define IVW_VTKUNSTRUCTUREDGRIDTORECTILINEARGRID_H
-
-#include <modules/tensorvisio/tensorvisiomoduledefine.h>
-#include <inviwo/core/common/inviwo.h>
-#include <inviwo/core/processors/processor.h>
-#include <inviwo/core/properties/fileproperty.h>
-#include <inviwo/core/properties/ordinalproperty.h>
-#include <inviwo/core/properties/buttonproperty.h>
-#include <inviwo/core/processors/progressbarowner.h>
-#include <inviwo/core/processors/activityindicator.h>
-#include <modules/tensorvisio/ports/vtkdatasetport.h>
-
-#include <atomic>
-
-#include <warn/push>
-#include <warn/ignore/all>
-#include <vtkSmartPointer.h>
-#include <vtkDataSet.h>
-#include <warn/pop>
-
-namespace inviwo {
-
-/** \docpage{org.inviwo.VTKUnstructuredGridToRectilinearGrid, VTK Unstructured Grid To Rectilinear
- * Grid}
- * ![](org.inviwo.VTKUnstructuredGridReader.png?classIdentifier=org.inviwo.VTKUnstructuredGridToRectilinearGrid)
- */
-
-class IVW_MODULE_TENSORVISIO_API VTKUnstructuredGridToRectilinearGrid
-    : public Processor,
-      public ActivityIndicatorOwner,
-      public ProgressBarOwner {
-public:
-    VTKUnstructuredGridToRectilinearGrid();
-    virtual ~VTKUnstructuredGridToRectilinearGrid();
-
-    virtual void process() override;
-
-    virtual const ProcessorInfo getProcessorInfo() const override;
-    static const ProcessorInfo processorInfo_;
-
-protected:
-    enum class State { NotStarted, Working, Abort, Done, Invalid };
-
-    struct WorkerState {
-        std::atomic<State> state = State::NotStarted;
-        std::atomic<bool> abortConversion_ = false;
-        std::atomic<bool> processorExists_ = true;
-
-        ivec3 dims_ = ivec3(1);
-    };
-
-    void loadData();
-
-    IntProperty maxDimension_;
-
-    ButtonProperty button_;
-    ButtonProperty abortButton_;
-
-    VTKDataSetInport inport_;
-    VTKDataSetOutport outport_;
-
-    std::shared_ptr<VTKDataSet> data_;
-    vtkSmartPointer<vtkDataSet> dataSet_;
-    std::shared_ptr<WorkerState> state_;
-};
-
-}  // namespace inviwo
-
-#endif  // IVW_VTKUNSTRUCTUREDGRIDTORECTILINEARGRID_H
